Update README.md
This program describes the implementation of a machine learning- based email classification system using Python. The system uses the Naive Bayes algorithm to determine whether an email is spam or not. The development process uses data from a set of emails to train the model, evaluate its quality, and predict the status of new emails.